web/requirement.txt
appnope==0.1.2
APScheduler==3.6.3
asgiref==3.4.0
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
decorator==5.0.9
Django==3.2.5
django-extensions==3.1.3
djangorestframework==3.12.4
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
python-telegram-bot==13.6
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.6
wcwidth==0.2.5
